Porting Mishap
Lots in here so far: all Tele, including single and dual controller, and one Auto. Camera.

None of it is tested, though it all compiles.

Most of it is a port that remains true to the original quirks/bugs. I cleaned up a few spots, but lots more to do if you were gonna take this fwd. The point of this is to get a good look at how the commands/subsystems work for our code this year, and to see what would be good to have common. Lots of hints so far.