Caching goodness
Get some bulk caching jank out of the MikeBot test code now that StealthOpMode does the manual caching setup and clearing automatically.

Previously I had to be aware of the bulk cache, that it was in AUTO mode, and that I had to clear it to make my arm work well in the face of other concurrent actions. All that crap is gone now!