Upgrade to v7.2 of FTC SDK
FIRST released v7.2 of the FTC SDK on 8/2/2022, updating their git repo https://github.com/FIRST-Tech-Challenge/FtcRobotController with a new version of the robot controller, and also releasing a v7.2 of the driver station APK on the releases page.

Updated this project with the 7.2 SDK changes. I did this by cloning the whole FtcRobotController repo, and copying everything (except the .git dir!!) on top of this repo. That did a couple of things. First, all the build files and changes to the FtcRobotController directory were made visible as changes in git. (Except for deleted files, which bit me in samples, but was easily corrected.) It also added the TeamCode directory, which I then copied over the StarterCode dir to also see changes.

Overall the changes are minor and could be merged easily and cleanly. I kept everything that we needed for adding FTC Lib and RoadRunner. I took all of their changes to the builds, removing old supported devices, upgrading gradle, etc. since those all make perfect sense.

This works on my test bot just fine, though I had to side-load the v7.2 DS on my Android phone. There is no 7.2 DS option available for the REV Driver Hubs yet, so I'm putthing this into a branch for now. Once we have updated driver hubs we can merge this into the main branch and proceed from there.