Changed auto-to-teleop heading storage
I had something that seemed simple to use, but doesn't actually work for simple bots. You can't query the IMU after the opmode ends, it returns 0 all the time.

Added an explicit command to save the last heading from auto, which we can place at the end of our auto sequences.

Added a command to stop the opmode, so we can end auto as soon as we're done and not have to wait for the 30s to elapse.